adds a couple of hooks to edit-with-emacs workflow (#129)
We already expose a hook for when edit-with-emacs starts, `spacehammer/edit-with-emacs-hook`

It gets called at the beginning of the edit and gives users a chance to set frame, buffer and other parameters, major and minor modes, e.g., one may want to set a "distraction free mode", or/and visual-line-mode, etc.

-----

These newly added hooks get called right before closing the edit frame and, for example, can be used like this:

    (defun spacehammer-before-finish-edit-with-emacs (bufname pid)
      (with-current-buffer bufname
        (set-buffer-modified-p nil)))

    (add-hook 'spacehammer/before-finish-edit-with-emacs-hook #'spacehammer-before-finish-edit-with-emacs)

In that particular snippet, we mark the buffer as unmodified, so Emacs wouldn't prompt, asking if the file needs saving. It doesn't usually do that (because Spacehammer, by default, doesn't designate a file). However, if in `spacehammer/edit-with-emacs-hook` we set a filename for the buffer, i.e.:

  (set-visited-file-name (format "/tmp/%s_%s_%s" buffer-name pid title))

Then, the previous snippet may be desired.

So you may ask: "why would you even want to designate a file, the text gets stored in the kill-ring anyway?". There could be multiple different reasons for that. The main one for me is the lsp-mode. The way how it works - it requires a buffer content to be bound to a file on disk.