apps: Update app even when new app isn't in config
We want to do this so that once we leave an app the state of the machine
has :app set to nil, otherwise the modal menu will continue to display
items from the previous app.

Because we are still firing an effect, we need to guard around the
effect handlers, doing nothing when we have an enter or launch effect
but no app.