remove needless coroutine (#187)
According the git-blame the couroutine was added to alleviate (perhaps some
imaginary problems), back in the day when "edit-with-emacs" was using a
separate frame. I think the garbage collector in some cases fails to
clean it up. I'm still not completely sure if that's the exact thing
that's causing [agzam/spacehammer#186 — Hammerspoon is eggrediously
unstable on macOS
Sonoma](https://github.com/agzam/spacehammer/issues/186), but after
removing the coroutine and testing for couple of days, I'm no longer
experiencing the issues described in #186