add stateful container for named loggers
State manager for loggers, by id. The primary usecase is for users to
override the log level of a given logger id using a central location to
store them by ID.

Table is weak-valued so loggers can be cleaned up when they have no
remaining references outside of the container. Invoking the table as a
function ala `(logger "logger-id" ?some-level)` allows users to fetch a
logger with a given ID if it's already been instantiated, otherwise
creating and storing a new one by passing arguments to `hs.logger.new`